mpi4py.futures: Rename internal module
