+++ src/mpi4py/futures/_base.pyi
