+++ src/mpi4py/futures/aplus.pyi
@@ -1,4 +1,4 @@
-from ._core import Future
+from ._base import Future
 from typing import Generic, Optional, TypeVar, Union
 from typing import Callable
